pyproject.toml
[tool.poetry]
name = "cmms"
version = "0.0.1"
description = "A reliable scheduling, tracking, reporting tools for equipment and facilities maintenance."
authors = [
  "FATx64",
  "null2264"
]
license = "BSD-3-Clause"

[tool.poetry.scripts]
start = "scripts:runserver"
tailwind = "scripts:tailwind"
manage = "scripts:manage"

[tool.poetry.dependencies]
python = "^3.10"
Django = {extras = ["bcrypt"], version = "^4.2.1"}
dj-database-url = "^2.0.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
django-widget-tweaks = "^1.4.12"
typing-extensions = "^4.5.0"
bcrypt = "^4.0.1"
django-redis = "^5.2.0"
pillow = "^9.5.0"
dateutils = "^0.6.12"

[tool.poetry.group.dev.dependencies]
django-tailwind = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 125
skip-string-normalization = true

[tool.isort]
profile = "black"
multi_line_output = 3
lines_after_imports = 2
